requirements.txt
bcrypt==4.2.1
blinker==1.9.0
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==42.0.8
dnspython==2.7.0
Flask==3.1.0
future==1.0.0
impacket==0.12.0
itsdangerous==2.2.0
Jinja2==3.1.4
keyboard==0.13.5
ldap3==2.9.1
ldapdomaindump==0.9.4
MarkupSafe==3.0.2
MouseInfo==0.1.3
numpy==2.1.3
opencv-python==4.10.0.84
paramiko==3.5.0
pillow==11.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pycparser==2.22
pycryptodomex==3.21.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyOpenSSL==24.0.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python3-xlib==0.15
pytweening==1.2.0
scp==0.15.0
setuptools==75.5.0
six==1.16.0
Werkzeug==3.1.3
